fix(Super-Candy): fix Super Candy speeds up the snake

+++ Commons/include/GameObject/IGameObject.hpp
@@ -19,7 +19,7 @@ class IGameObject {
 
         struct Effects {
             int deltaScore;
-            int deltaSpeed;
+            float deltaSpeed;
             int deltaGrowth;
             bool killPlayer;
         };
